Finish optimizing and refactoring for bckt model
